Pausing & Refactoring
# Description
Players are now able to pause locally as part of the functionality of `ANetworkFYPPlayerState`. This includes the option to finally close the game inside the game in the pause `UserWidget`.

To more easily get access to the `PlayerState` once it becomes valid, as usually required by the **UI**, attempting to receiving a call back once the `PlayerState` becomes valid has been exposed to BPs for reusability *(used to be implemented as part of score `UserWidget` blueprint)*.

Updated score `UserWidget` blueprint to use the new functionality to receive a call back with valid player state.

`HandleFire` method was renamed to include the "Server" prefix to better align with Unreal's naming convention.

# Known Issues
- Still unable to close the game until joining the game level.
- Decided to add the "b" prefix to `bool` type names as well to continue following naming conventions, however I have yet to update all `bool` type names. I'll change them as I come across them.