Online Service EOS Exploration
# Description
Using the knowledge about the EOS subsystem and the documentation available for it, I briefly explored how using the Online Service EOS plugin would look like. I managed to implement all the way up to creating a lobby session, at which I encountered an error with invalid state. However I could not find any documentation detailing the error.

All code related to the subsystem was commented out instead of removed to keep as a comparison point.

## NetworkFYPOnlineServiceSubsystem
Following documentation about using the EOS, I created a subsystem to manage all EOS related interactions.

Login and finding sessions were successfully implemented. The rest of the functions could not be tested due to not being able to create a lobby to test the remaining functions. Dedicated servers were not explored.